refactors characters and adds enemy to playfield
